Move stories to broken directory
